Move 'Use Newest App Version From HQ' to normal preferences
